localstack/requirements.txt
boto3
localstack-client
requests
PyJWT
huggingface_hub